requirements.txt
pytest==9.0.1
tqdm==4.66.5
pyyaml==6.0.2
requests==2.32.3
google-generativeai==0.8.3
python-dotenv[cli]==1.0.1

